+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/breakingnews/BreakingNewsPatch.kt
@@ -2,11 +2,11 @@ package app.revanced.patches.youtube.layout.hide.breakingnews
 
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.patch.BytecodePatch
-import app.revanced.patches.youtube.layout.hide.suggestionsshelves.HideSuggestionsShelvesPatch
+import app.revanced.patches.youtube.layout.hide.shelves.HideHorizontalShelvesPatch
 
-@Deprecated("This patch class has been renamed to HideSuggestionsShelvesPatch.")
+@Deprecated("This patch class has been renamed to HideHorizontalShelvesPatch.")
 object BreakingNewsPatch : BytecodePatch(
-    dependencies = setOf(HideSuggestionsShelvesPatch::class),
+    dependencies = setOf(HideHorizontalShelvesPatch::class),
 ) {
     override fun execute(context: BytecodeContext) {
     }
