+++ src/main/resources/addresources/values/strings.xml
@@ -484,10 +484,10 @@
             <string name="revanced_hide_album_cards_summary_on">Album cards are hidden</string>
             <string name="revanced_hide_album_cards_summary_off">Album cards are shown</string>
         </patch>
-        <patch id="layout.hide.suggestionsshelves.HideSuggestionsShelvesPatch">
-            <string name="revanced_hide_suggestions_shelves_title">Hide suggestions shelves</string>
-            <string name="revanced_hide_suggestions_shelves_summary_on">Shelves are hidden such as:\n• Breaking news\n• Continue watching\n• Explore more channels\n• Shopping\n• Watch it Again</string>
-            <string name="revanced_hide_suggestions_shelves_summary_off">Suggestions shelves are shown</string>
+        <patch id="layout.hide.shelves.HideHorizontalShelvesPatch">
+            <string name="revanced_hide_horizontal_shelves_title">Hide horizontal shelves</string>
+            <string name="revanced_hide_horizontal_shelves_summary_on">Shelves are hidden such as:\n• Breaking news\n• Continue watching\n• Explore more channels\n• Shopping\n• Watch it Again</string>
+            <string name="revanced_hide_horizontal_shelves_summary_off">Shelves are shown</string>
         </patch>
         <patch id="layout.hide.comments.CommentsPatch">
             <string name="revanced_comments_screen_title">Comments</string>
